+++ vertigo-ui/src/main/resources/io/vertigo/ui/static/js/vertigo-ui.umd.js
@@ -8208,7 +8208,7 @@ function isString (v) {
     }
   },
   uploader_uploadedFiles: function uploader_uploadedFiles(uploadInfo, myComponentId, fileInfoKey) {
-    uploadInfo.files.forEach(function (file, index, array) {
+    uploadInfo.files.forEach(function (file) {
       var response = JSON.parse(file.xhr.response);
       this.$data.vueData.CTX = response.model.CTX;
       Object.keys(response.model).forEach(function (key) {
@@ -8229,7 +8229,16 @@ function isString (v) {
       Object.keys(response.uiMessageStack).forEach(function (key) {
         this.$data.uiMessageStack[key] = response.uiMessageStack[key];
       }.bind(this));
-      array.splice(index, 1);
+    }.bind(this));
+  },
+  uploader_failedFiles: function uploader_failedFiles(uploadInfo) {
+    uploadInfo.files.forEach(function (file) {
+      var response = JSON.parse(file.xhr.response); //server can return : a response with a uiMessageStack object or directly the uiMessageStack
+
+      var uiMessageStack = response.globalErrors ? response : response.uiMessageStack;
+      Object.keys(uiMessageStack).forEach(function (key) {
+        this.$data.uiMessageStack[key] = uiMessageStack[key];
+      }.bind(this));
     }.bind(this));
   },
   uploader_removeFiles: function uploader_removeFiles(removedFiles, componentId
@@ -8238,29 +8247,33 @@ function isString (v) {
     var component = this.$refs[componentId];
     removedFiles.forEach(function (removedFile) {
       var xhrParams = {};
-      xhrParams[component.fieldName] = removedFile.fileUri;
-      xhrParams['CTX'] = this.$data.vueData.CTX;
-      this.$http.delete(component.url, {
-        params: xhrParams,
-        credentials: component.withCredentials
-      }).then(function (response) {
-        //Ok
-        if (response.data.model.CTX) {
-          this.$data.vueData.CTX = response.data.model.CTX;
-        }
 
-        Object.keys(response.data.model).forEach(function (key) {
-          if ('CTX' != key) {
-            this.$data.vueData[key] = response.data.model[key];
+      if (removedFile.fileUri) {
+        //if file is serverside
+        xhrParams[component.fieldName] = removedFile.fileUri;
+        xhrParams['CTX'] = this.$data.vueData.CTX;
+        this.$http.delete(component.url, {
+          params: xhrParams,
+          credentials: component.withCredentials
+        }).then(function (response) {
+          //Ok
+          if (response.data.model.CTX) {
+            this.$data.vueData.CTX = response.data.model.CTX;
           }
+
+          Object.keys(response.data.model).forEach(function (key) {
+            if ('CTX' != key) {
+              this.$data.vueData[key] = response.data.model[key];
+            }
+          }.bind(this));
+          Object.keys(response.data.uiMessageStack).forEach(function (key) {
+            this.$data.uiMessageStack[key] = response.data.uiMessageStack[key];
+          }.bind(this));
+        }.bind(this)).catch(function (error) {
+          //Ko
+          this.$q.notify(error.response.status + ":" + error.response.statusText + " Can't remove temporary file");
         }.bind(this));
-        Object.keys(response.data.uiMessageStack).forEach(function (key) {
-          this.$data.uiMessageStack[key] = response.data.uiMessageStack[key];
-        }.bind(this));
-      }.bind(this)).catch(function (error) {
-        //Ko
-        this.$q.notify(error.response.status + ":" + error.response.statusText + " Can't remove temporary file");
-      }.bind(this));
+      }
     }.bind(this));
     this.uploader_forceComputeUploadedSize(componentId);
   },
